feat: Demo the same json->struct example using Nim std libs
In this commit, the Nim std libs json and jsonutils are used to do the
same JSON string -> Nim object mapping done using the jsony library in
the previous commit.

The json+jsonutils combo is a bit more picky about the matching of the
Nim object element names and JSON key names, but it eventually works
the same way.

- To run the example using the std libs, run "make".
- To run the example using jsony, run "make jsony".